# docs/swagger.yaml
basePath: /
definitions:
...
  api.EnvironmentState:
    properties:
      as_of:
        description: |-
          AsOf is when the values were read, RFC3339. It is not decoration: a zone
          value with a time constant is on its way to a set point and is resolved to
          exactly this instant, so the number means nothing without it.
        example: "2026-08-27T09:41:00Z"
        type: string
      assets:
        additionalProperties:
          additionalProperties: true
          type: object
        type: object
      context:
        additionalProperties: true
        type: object
      history_running:
        description: |-
          HistoryRunning says that the environment stands at a past instant while a
          history run rebuilds it. It carries no state then and is not running in the
          live sense, which is why it is a flag of its own rather than a state an
          editor would show as current.
        example: false
        type: boolean
      running:
        description: |-
          Running says whether this instance simulates the environment at all. An
          environment that is stored but not running here is not an error - it is
          the normal case for a document that was just written, or one another
          instance runs - and it carries no state, which is what an editor turns
          into its empty state.
        example: true
        type: boolean
      zones:
        additionalProperties:
          additionalProperties: true
          type: object
        type: object
    type: object
  api.HistoryRequest:
    properties:
      force:
        description: |-
          Force starts the run although the first day of the window already holds
          readings for a device of this environment. Those rows stay and the run
          writes its own beside them; without it such a request is answered with
          409.
        example: false
        type: boolean
      from:
        example: "2026-07-01T00:00:00Z"
        type: string
    type: object
...
  repo.StateChange:
    properties:
      assets:
        additionalProperties:
          additionalProperties: true
          type: object
        type: object
      context:
        additionalProperties: true
        type: object
      zones:
        additionalProperties:
          additionalProperties: true
          type: object
        type: object
    type: object
...
  runtime.HistoryChannelStatus:
    properties:
      asset_id:
        type: string
      channel_id:
        type: string
      failed:
        type: integer
      last_error:
        type: string
      name:
        type: string
      publishable:
        type: boolean
      published:
        type: integer
      reason:
        type: string
      silent:
        type: integer
    type: object
  runtime.HistoryState:
    enum:
    - running
    - done
    - failed
    - cancelled
    type: string
    x-enum-varnames:
    - HistoryRunning
    - HistoryDone
    - HistoryFailed
    - HistoryCancelled
  runtime.HistoryStatus:
    properties:
      channels:
        description: |-
          Channels is what became of every channel the run drove. It is filled when
          the run ends and stays empty for one that broke, since a run that panicked
          returns nothing. It is what tells "the environment publishes nothing" from
          "every channel of it was unpublishable", which look identical in the
          totals.
        items:
          $ref: '#/definitions/runtime.HistoryChannelStatus'
        type: array
      environment_id:
        type: string
      error:
        type: string
      failed:
        type: integer
      finished_at:
        type: string
      from:
        type: string
      last_error:
        description: |-
          LastError is the most recent publish failure of the run, kept so a run that
          mostly worked still says what went wrong. Error is what broke the run
          itself.
        type: string
      position:
        description: Position is the virtual instant the run has reached.
        type: string
      published:
        description: |-
          Published and Failed count the steps of the publish grids of every channel:
          the ones that sent at least one reading, and the ones the platform refused.
        type: integer
      started_at:
        type: string
      state:
        $ref: '#/definitions/runtime.HistoryState'
      to:
        description: |-
          To is where the run ended, which is the present at the moment it stopped
          simulating: it moves forward past the instant of the request while the run
          chases the time it spent simulating.
        type: string
    type: object
info:
  contact: {}
  description: |-
    Simulates environments — sites, buildings and apartments — and
    their devices, and publishes their data as if it came from real
    hardware. An environment is one document: GET returns what PUT
    accepts, so a whole site can be created in a single call.

    The legacy world/room/device endpoints are being replaced by the
    environment api and are not documented here.
  title: MOSES API
  version: 1.0.0
paths:
...
  /environments/{id}/history:
    delete:
      description: |-
        Ends a running simulation of the past. It does not wait and it does not undo: the run stops at its next step and hands the environment back to the live simulation, which then continues from the partial state — a consistent state of an earlier instant. The response is the run as it stood when the abort was accepted.

        A run over a year of minute data suspends the live simulation for a long time, which is why it has a switch a backfill does not need.
      parameters:
      - description: environment id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: the abort was accepted
          schema:
            $ref: '#/definitions/runtime.HistoryStatus'
        "401":
          description: the token carries no subject
          schema:
            type: string
        "404":
          description: no such environment, no access to it, or nothing is known about
            a history run of it
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      security:
      - Bearer: []
      summary: Abort the history run of one environment
      tags:
      - Environment
    get:
      description: |-
        Where the run stands: running, done, failed or cancelled, which virtual instant it has reached, how many publish steps went out or were refused, and per channel what became of it — including the reason a channel published nothing at all. `done` means the live simulation is running again on the state the run arrived at; `failed` and `cancelled` mean it is running again on the partial state the run had reached, which is a consistent state of an earlier instant and not a rollback.

        A run is stored, so a restart does not forget it: a run that was still going is resumed from its last checkpoint - written at every virtual hour - and continues to be followed here, and the outcome of a run that is over stays readable. 404 therefore means that no run of this environment is known at all.
      parameters:
      - description: environment id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/runtime.HistoryStatus'
        "401":
          description: the token carries no subject
          schema:
            type: string
        "404":
          description: no such environment, no access to it, or nothing is known about
            a history run of it
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      security:
      - Bearer: []
      summary: The history run of one environment
      tags:
      - Environment
    post:
      consumes:
      - application/json
      description: |-
        Simulates the environment from `from` to the present on a virtual clock and publishes every reading under the instant it was computed for. Unlike a backfill the run carries state, and the state it arrives at **becomes** the live state: a meter reading continues into the live one as one ramp rather than starting a second one next to it. Runs asynchronously; the response is the run as it stands, `GET` follows it and `DELETE` aborts it.

        **The live simulation of this environment is suspended for the duration of the run and its current state is discarded.** While it runs, `PATCH /environments/{id}/state`, the state snapshot and a second run are answered with 409, incoming commands are dropped, and an edit to the definition takes effect when the run ends. State `done` means the live simulation is running again on the state the run arrived at.

        Only a channel whose platform service declares the attribute `senergy/time_path` can publish with a historical timestamp; every other channel is computed but sends nothing, so the environment as a whole still arrives at the right state. Every source kind is simulated, script and schedule included, which is what a backfill cannot do. Every channel that publishes nothing says why in `channels` of the status.

        There is no end: the run ends at the present, and once it has simulated the window it keeps going until it has caught up with the time it spent doing so, so that no gap is left at the handover. `to` in the status is therefore where the run actually ended.

        The window may not start in the future, has to be at least a minute and at most 366 days long, and may not come to more than twenty million simulation steps across the channels and context sources of the environment.

        **A window whose first day already holds readings is refused with 409**, naming the devices it found them for: that is either an earlier run over the same window or a window that reaches into live data, and the rows of a run cannot be deleted again. `force: true` starts the run anyway, which writes the readings of the window a second time. Only the first day from `from` is looked at, since the end of any window of an environment that has been running always holds readings. Without a configured timescale-wrapper the check is skipped. The queries run on the caller's token, and a channel it is answered a 404 for is asked again with the owner's; a channel left with a 400 or a 404 is unchecked with a warning, while every other answer - another status, an unavailable wrapper, a network error - refuses the run rather than starting it unchecked. A check that does not answer within five seconds is answered with 503.
      parameters:
      - description: environment id
        in: path
        name: id
        required: true
        type: string
      - description: the instant to start from
        in: body
        name: window
        required: true
        schema:
          $ref: '#/definitions/api.HistoryRequest'
      produces:
      - application/json
      responses:
        "202":
          description: accepted, with the run as it stands
          schema:
            $ref: '#/definitions/runtime.HistoryStatus'
        "400":
          description: the body is unreadable, or the window is in the future, too
            long or too dense
          schema:
            type: string
        "401":
          description: the token carries no subject
          schema:
            type: string
        "404":
          description: no such environment, no access to it, or it is not running
            here
          schema:
            type: string
        "409":
          description: a history run or a backfill of this environment is already
            running, or the first day of the window already holds readings and force
            is not set
          schema:
            type: string
        "413":
          description: the request body is larger than the allowed limit
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
        "503":
          description: 'the timescale did not answer in time, so the window could
            not be checked; retry or send force: true'
          schema:
            type: string
      security:
      - Bearer: []
      summary: Run one environment from a past instant up to now
      tags:
      - Environment
...
  /environments/{id}/state:
    get:
      description: |-
        What the running simulation currently holds: the shared context, the state of every zone and the state of every asset. This is the reading direction of PATCH on the same path and answers in the same shape, so a value can be read, changed and sent straight back.

        A zone value that the definition gives a time constant is on its way to its set point rather than at it; it is resolved to the instant in `as_of`, which is the same value a script would read at that moment.

        These are live values, not the definition: they are not what GET /environments/{id} returns, and they are not stored by reading them.

        An environment that is stored but not simulated here answers 200 with `running: false` and no states. That is not an error: another instance may run it, or it may just have been written. Only an environment that does not exist, or one the caller may not see, is a 404.

        While a history run rebuilds the environment from a past instant the answer is 200 with `running: false` and `history_running: true`. There is no live state to read then: the environment stands in the past, and reading it would resolve its values against the wall clock and corrupt the run.
      parameters:
      - description: environment id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.EnvironmentState'
        "401":
          description: the token carries no subject
          schema:
            type: string
        "404":
          description: no such environment, or no access to it
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      security:
      - Bearer: []
      summary: Live state of one environment
      tags:
      - Environment
    patch:
      consumes:
      - application/json
      description: |-
        Merges the given values into the live state. Only the keys named are touched; everything else keeps running. Context is the shared surroundings every zone reads, zones and assets are keyed by their id from the definition. Takes effect on the next tick of the scripts that read it, and is not a change to the definition.

        A context key the environment's `timeline` governs is a declared function of time and cannot be moved from here: a value differing from the declared one is refused with 400. Sending the value GET on this path handed out is not a move and is accepted, so reading the state, editing one key and sending the whole thing back works unchanged.
      parameters:
      - description: environment id
        in: path
        name: id
        required: true
        type: string
      - description: the values to set
        in: body
        name: state
        required: true
        schema:
          $ref: '#/definitions/repo.StateChange'
      produces:
      - application/json
      responses:
        "204":
          description: applied
          schema:
            type: string
        "400":
          description: the body is unreadable, empty, holds a value that is not plain
            data or nests deeper than 32 levels or holds more than 10000 elements,
            names a zone or asset the definition does not have, or moves a context
            key the timeline governs; a change wrong in both ways names both
          schema:
            type: string
        "401":
          description: the token carries no subject
          schema:
            type: string
        "404":
          description: no such environment, no access to it, or it is not running
            here
          schema:
            type: string
        "409":
          description: a history run of this environment is in progress, so it stands
            at a past instant
          schema:
            type: string
        "413":
          description: the request body is larger than the allowed limit
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      security:
      - Bearer: []
      summary: Set live state of one running environment
      tags:
      - Environment
securityDefinitions:
  Bearer:
    description: A keycloak issued JWT. Verified at the gateway, not here.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"